+++ app/Http/Controllers/DocumentoController.php
@@ -7,6 +7,7 @@ use ArqAdmin\Repositories\DocumentoRepository;
 use ArqAdmin\Services\DocumentoService;
 use Illuminate\Http\Request;
 
+
 class DocumentoController extends Controller
 {
     /**
@@ -50,6 +51,8 @@ class DocumentoController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('role-documental');
+
         return $this->service->create($request->all());
     }
 
@@ -73,6 +76,8 @@ class DocumentoController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('role-documental');
+
         return $this->service->update($request->all(), $id);
     }
 
@@ -84,6 +89,8 @@ class DocumentoController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('role-documental');
+
         return $this->service->delete($id);
     }
 
@@ -94,6 +101,8 @@ class DocumentoController extends Controller
 
     public function getRevisionHistory($id)
     {
+        $this->authorize('role-documental');
+
         return $this->service->getRevisionHistory($id);
     }
 }
